requirements.txt
# ============================================================
# Telegram Framework
# ============================================================
pyrogram==2.0.106
tgcrypto==1.2.5
pytz

# ============================================================
# Video Metadata & Thumbnails
# ============================================================
hachoir
Pillow

# ============================================================
# Database Connection
# ============================================================
pymongo
dnspython

# ============================================================
# Asynchronous Operations
# ============================================================
aiofiles
httpx

# ============================================================
# Environment Variable Loader
# ============================================================
python-dotenv
